pyproject.toml
[project]
name = "whos-bailey"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask[async]>=3.1.1",
    "flask-cors>=5.0.0",
    "google-adk>=1.7.0",
]
